externalize workerd logic into its own file
